+++ mrv2/lib/mrvCore/mrvHotkey.cpp
@@ -63,6 +63,7 @@ namespace mrv
     Hotkey kSafeAreas(false, false, false, false, 's');
     Hotkey kIgnoreDisplayWindow(false, false, false, false, 0);
     Hotkey kAutoNormalize(false, false, false, false, 0);
+    Hotkey kInvalidValues(false, false, false, false, 0);
     
     Hotkey kDisplayWindow(true, false, false, false, 'd');
     Hotkey kDataWindow(false, false, false, false, 'd');
@@ -384,6 +385,7 @@ namespace mrv
         HotkeyEntry(_("Safe Areas"), &kSafeAreas),
         HotkeyEntry(_("Ignore Display Window"), &kIgnoreDisplayWindow),
         HotkeyEntry(_("Auto Normalize"), &kAutoNormalize),
+        HotkeyEntry(_("Invalid Values"), &kInvalidValues),
         HotkeyEntry(_("Display Window"), &kDisplayWindow),
         HotkeyEntry(_("Data Window"), &kDataWindow),
         HotkeyEntry(_("Compare Wipe"), &kCompareWipe),