+++ mrv2/lib/mrvFl/mrvCallbacks.h
@@ -114,9 +114,12 @@ namespace mrv
     //! Ignore Display Window callback
     void ignore_display_window_cb(Fl_Menu_* w, ViewerUI* ui);
     
-    //! normalize callback
+    //! Normalize callback
     void normalize_image_cb(Fl_Menu_* w, ViewerUI* ui);
 
+    //! Invalid values callback
+    void invalid_values_cb(Fl_Menu_* w, ViewerUI* ui);
+    
     //! Masking callback
     void masking_cb(Fl_Menu_* w, ViewerUI* ui);