+++ mrv2/docs/HISTORY.md
@@ -2,6 +2,8 @@ v1.2.3
 ======
 
 - Fixed loading of movies with multiple audio tracks.
+- Added "View/Auto Normalize" to normalize HDR (OpenEXR and HDR images).
+- Added "View/Invalid Values" to show HDR images' invalid values.
 
 
 v1.2.2